move dynamic scaling rope to intermlm2.py
